API 에러로그 추가 uptime,savingtime int -> long

+++ hans-sses/src/main/java/com/hans/sses/cms/admin/controller/DashBoardController.java
@@ -138,9 +138,9 @@ public class DashBoardController {
 			
 			for(int j=0;j<uptime.length;j++){
 				
-				int w = Integer.parseInt(watt[j]);
-				int u = Integer.parseInt(uptime[j]);
-				int s = Integer.parseInt(savingtime[j]);
+				long w = Long.parseLong(watt[j]);
+				long u = Long.parseLong(uptime[j]);
+				long s = Long.parseLong(savingtime[j]);
 				
 				
 				if(type.equals("on")){
